add more stats
